# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/reservations":{"put":{"tags":["reservation-controller"],"operationId":"checkInReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["reservation-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}}},"/flights":{"post":{"tags":["flight-controller"],"operationId":"saveFlight","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Flight"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Flight"}}}}}}},"/reservations/{id}":{"get":{"tags":["reservation-controller"],"operationId":"getReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/flights/{id}":{"get":{"tags":["flight-controller"],"operationId":"getFlightById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Flight"}}}}}}}},"components":{"schemas":{"PassengetDto":{"type":"object","properties":{"name":{"type":"string"},"seat":{"type":"string"}}},"ReservationDto":{"type":"object","properties":{"id":{"type":"string"},"flight":{"type":"string"},"status":{"type":"string"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/PassengetDto"}}}},"Airport":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"code":{"type":"string"},"location":{"type":"string"}}},"Flight":{"type":"object","properties":{"id":{"type":"string"},"aircraft":{"type":"string"},"departure":{"type":"string","format":"date-time"},"airport":{"$ref":"#/components/schemas/Airport"},"seat_map":{"type":"array","items":{"$ref":"#/components/schemas/Seat"}},"reservations":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}},"Passenger":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"reservation":{"$ref":"#/components/schemas/Reservation"}}},"Reservation":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string"},"flight":{"$ref":"#/components/schemas/Flight"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/Passenger"}}}},"Seat":{"type":"object","properties":{"seat":{"type":"string"},"status":{"type":"string"},"flight":{"$ref":"#/components/schemas/Flight"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}