# src/main/java/ar/edu/utn/frc/tup/lc/iv/models/PassengetDto.java
package ar.edu.utn.frc.tup.lc.iv.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.RequiredArgsConstructor;

@Data
@RequiredArgsConstructor
@AllArgsConstructor
public class PassengetDto {
    public String name;
    public String seat;

}